Corrected calculations for power and heating

+++ src/main/java/pl/put/poznan/building/classes/Room.java
@@ -36,7 +36,7 @@ public class Room extends Location {
     }
 
     @Override
-    public int getLightingPower() {
+    public float getLightingPower() {
         return lighting;
     }
